QTZ-326: Ensure we re-throw exception if it's not the case of no class found.

+++ quartz-core/src/main/java/org/quartz/xml/XMLSchedulingDataProcessor.java
@@ -1009,12 +1009,12 @@ public class XMLSchedulingDataProcessor implements ErrorHandler {
                 // loading the job class, so use try/catch to handle it.
                 dupeJ = sched.getJobDetail(detail.getKey());
             } catch (JobPersistenceException e) {
-                if (e.getCause() instanceof ClassNotFoundException) {
-                    if (isOverWriteExistingData()) {
-                        // We are going to replace jobDetail anyway, so just delete it first.
-                        log.info("Removing job: " + detail.getKey());
-                        sched.deleteJob(detail.getKey());
-                    }
+                if (e.getCause() instanceof ClassNotFoundException && isOverWriteExistingData()) {
+                    // We are going to replace jobDetail anyway, so just delete it first.
+                    log.info("Removing job: " + detail.getKey());
+                    sched.deleteJob(detail.getKey());
+                } else {
+                    throw e;
                 }
             }
